+++ src/main/java/lotto/domains/lotto/domain/Lotto.java
@@ -1,4 +1,4 @@
-package lotto.domains.lotto;
+package lotto.domains.lotto.domain;
 
 import java.util.HashSet;
 import java.util.List;
